Align branches with latest Experience, Education, and Employment models. JobApplication, JobDescription, and Document models aligned with structure of models and implemented.
- JobAppController implementation is added in this PR.
- The import logic has been changed to be more explicit, rather than importing all the files in /models/, the specific models which need to be sequelized by the db have been specified explicitly. There is tighter coupling here but the explicit declaration is better for cross-platform compatibility. Windows machine was having issues sequelizing database due previous /models/index.js implementing file pathing to search for models to initialize.

Next step:
- Implement and test Experience models, ExperienceController, Job Application models, and JobApplicationController.
