+++ src/controllers/index.js
@@ -1,5 +1,5 @@
-import User from './User.js';
-import Experience from './Experience.js';
+import User from './user.js';
+import Experience from './experience.js';
 
 export default {
   User,
